chore: renaming of event dispatcher. moving events fetcher to its own package
